pyproject.toml
[tool.poetry]
name = "mango-trading"
version = "0.1.0"
description = ""
authors = ["corn <dalcon10280@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "^3.11.11"
pydantic-settings = "^2.7.0"
pytest = "^8.3.4"
pendulum = "^3.0.0"
pandas = "^2.2.3"
numpy = "^2.2.1"
backtesting = "^0.3.3"
jupyter = "^1.1.1"
websockets = "^14.1"
python-binance = "^1.0.26"
pip-system-certs = "^4.0"
yarl = "^1.18.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-aio = "^1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[market-data.binance]
market-data-endpoint = "https://data-api.binance.vision/api/v3"